github-scraper/requirements.txt
requests>=2.28.1
beautifulsoup4>=4.11.1
lxml>=4.9.1
pandas>=1.5.1
PyGithub>=1.58.1
tqdm>=4.64.1
python-dotenv>=0.21.0
openpyxl>=3.0.10
argparse>=1.4.0
pytest>=7.2.0
flake8>=6.0.0
black>=22.12.0
flask>=2.2.3
flask-cors>=3.0.10 